Move the shared exception handler to the Bit.ExceptionHandling namespace
The filter and its WithBasicExceptionHandling extension were left in
Bit.Core.Utilities when they were extracted into src/Libraries/ExceptionHandling,
which no longer matches where they live. Namespace them after the library, as the
sibling HttpExtensions library does with Bit.HttpExtensions.

The exception types and ErrorResponseModel keep their Bit.Core.Exceptions and
Bit.Core.Models.Api namespaces: they have call sites throughout the codebase, so
moving them is a separate, much wider change.
